+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # 2.x branch
+## 2.16 branch
+### 2.16.0
+* renamed `/ContactForm.php` as `ContactUsForm` and `ContactFormMailer` as
+    `ContactUsMailer`. This creates less confusion with the classes names.
+
 ## 2.15 branch
 ### 2.15.3
 * fixed code for `BackupForm` class. Added `_getBackupExportInstance()` method;